+++ internal/postgresql/postgresql.go
@@ -567,11 +567,11 @@ func (p *Manager) SetupRoles() error {
 	if p.suUsername == p.replUsername {
 		log.Infow("adding replication role to superuser")
 		if p.suAuthMethod == "trust" {
-			if err := alterPasswordlessRole(ctx, p.localConnParams, []string{"replication"}, p.suUsername); err != nil {
+			if err := alterPasswordlessRole(ctx, p.localConnParams, p.suUsername); err != nil {
 				return fmt.Errorf("error adding replication role to superuser: %v", err)
 			}
 		} else {
-			if err := alterRole(ctx, p.localConnParams, []string{"replication"}, p.suUsername, p.suPassword); err != nil {
+			if err := alterRole(ctx, p.localConnParams, p.suUsername, p.suPassword); err != nil {
 				return fmt.Errorf("error adding replication role to superuser: %v", err)
 			}
 		}
@@ -585,14 +585,13 @@ func (p *Manager) SetupRoles() error {
 			}
 			log.Infow("superuser password set")
 		}
-		roles := []string{"login", "replication"}
 		log.Infow("creating replication role")
 		if p.replAuthMethod != "trust" {
-			if err := createRole(ctx, p.localConnParams, roles, p.replUsername, p.replPassword); err != nil {
+			if err := createRole(ctx, p.localConnParams, p.replUsername, p.replPassword); err != nil {
 				return fmt.Errorf("error creating replication role: %v", err)
 			}
 		} else {
-			if err := createPasswordlessRole(ctx, p.localConnParams, roles, p.replUsername); err != nil {
+			if err := createPasswordlessRole(ctx, p.localConnParams, p.replUsername); err != nil {
 				return fmt.Errorf("error creating replication role: %v", err)
 			}
 		}